Zabezpieczenie skryptu chce_postgres_backup

+++ scripts/chce_postgres_backup.sh
@@ -15,7 +15,7 @@ DB_USER="postgres"
 DB_PASSWORD="xxx"
 
 #Numbers of days you want to keep copy of your databases
-NUMBER_OF_DAYS=180
+NUMBER_OF_DAYS=90
 echo "Dumping database to ${BACKUP_DIR}${DB_NAME}_${BACKUP_DATE}.sql" 
 
 if [ ! -d "${BACKUP_DIR}" ]; then
@@ -23,7 +23,11 @@ if [ ! -d "${BACKUP_DIR}" ]; then
 fi 
    
 PGPASSWORD="${DB_PASSWORD}" pg_dump -h "${DB_HOST}" -p "${DB_PORT}" -U "${DB_USER}" -d "${DB_NAME}" -w --format=custom | xz > "${BACKUP_DIR}${DB_NAME}_${BACKUP_DATE}.xz"
-
-echo "Dumping database finnish" 
-
+pg_dump_check=$?
+if [ "$pg_dump_check" - eq 0 ]; then
+echo "Dumping database completed successfully" 
 find "${BACKUP_DIR}" -type f -prune -mtime +"${NUMBER_OF_DAYS}" -exec rm -f {} \;
+else
+echo "Dumping database ended with an error"
+fi
+
